asr/requirements.txt
fastapi
uvicorn[standard]
openai-whisper
torch
torchaudio
python-multipart
